+++ Start/views/SelectActionView.h
@@ -31,7 +31,7 @@
 - (void) quickSelectCell;
 - (void) selectActionWithTitle:(NSString *)searchTitle;
 
-- (int) actionIDWithTitle:(NSString *)searchTitle;
+- (NSInteger) actionIDWithTitle:(NSString *)searchTitle;
 - (NSString *)actionTitleWithID:(int)theID;
 
 - (id) initWithFrame:(CGRect)frame delegate:(id<SelectActionViewDelegate>)aDelegate actions:(NSArray *)theActions;
